init project

+++ database/migrations/2023_12_22_054621_create_repositories_table.php
@@ -17,9 +17,9 @@ return new class extends Migration {
             $table->string('name', 200);
             $table->unsignedBigInteger('agency_id')->nullable();
             $table->foreign('agency_id')->references('id')->on('agencies')->onDelete('no action');
-            $table->unsignedInteger('province_id')->nullable();
+            $table->unsignedSmallInteger('province_id')->nullable();
             $table->foreign('province_id')->references('id')->on('provinces')->onDelete('no action');
-            $table->unsignedInteger('county_id')->nullable();
+            $table->unsignedSmallInteger('county_id')->nullable();
             $table->foreign('county_id')->references('id')->on('provinces')->onDelete('no action');
             $table->string('address', 2048)->nullable();
             $table->timestamps();
